Move responseText to ViewStateModel
