+++ src/main/java/com/notably/model/suggestion/SuggestionModel.java
@@ -1,9 +1,7 @@
 package com.notably.model.suggestion;
 
 import java.util.List;
-import java.util.Optional;
 
-import javafx.beans.property.Property;
 import javafx.collections.ObservableList;
 
 /**
@@ -22,25 +20,6 @@ public interface SuggestionModel {
      * @param suggestions The list of paths or notes that match the user input.
      */
     void setSuggestions(List<SuggestionItem> suggestions);
-
-    /**
-     * Shows user the meaning of their input, e.g. "Create a new note".
-     * When input cannot get corrected, the "error" message will be shown in commandTextProperty.
-     * TODO: add logo.
-     * @return The meaning of the user input.
-     */
-    Property<Optional<String>> responseTextProperty();
-
-    /**
-     * Updates the responseTextProperty.
-     * @param responseText The meaning of the "command" that the user inputs.
-     */
-    void setResponseText(String responseText);
-
-    /**
-     * Resets the command text when the user clears the input text field.
-     */
-    void clearResponseText();
 
     /**
      * Resets the list of suggestions.
